Move rdmPoToJson from scripts to webpack
