+++ webpack.common.config.js
@@ -1,4 +1,5 @@
 var webpack = require('webpack');
+var gettextParser = require('gettext-parser');
 var path = require('path');
 var fs = require('fs');
 
@@ -131,6 +132,26 @@ addons.addons.forEach(function(addonName) {
 
 fs.writeFileSync(staticPath('js/_allLogTexts.json'), JSON.stringify(mainLogs));
 fs.writeFileSync(staticPath('js/_anonymousLogTexts.json'), JSON.stringify(anonymousLogs));
+
+function rdmPoToJson() {
+    var acceptLanguages = ['en','ja'];
+    var translationsBaseDir = 'translations';
+    var getTextDomain = 'js_messages';
+    var poRelativePath = path.resolve(__dirname, 'website', translationsBaseDir) + '/';
+    var jsonRelativePath = path.resolve(__dirname, 'website', 'static', 'js', translationsBaseDir) + '/';
+    var localeDir = 'LC_MESSAGES';
+    var langCode;
+    var input;
+    var po;
+    for(var i=0 ; i<acceptLanguages.length ; i++){
+        langCode = acceptLanguages[i];
+        input = fs.readFileSync(poRelativePath + langCode + '/' + localeDir + '/' + getTextDomain + '.po');
+        po = gettextParser.po.parse(input);
+        fs.writeFileSync(jsonRelativePath + langCode + '.json' , JSON.stringify(po));
+    }
+}
+
+rdmPoToJson();
 
 var resolve = {
     modules: [
